# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Local Community Back End Demonstration API
  description: A Local Community API, used to register user, create new issue, organise meetings and deliver announcements
  contact:
    name: Harman Singh
    email: singhh48@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://outsidefilter-epoxyderby-8080.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /announcements:
    summary:
    description:

    summary: Rappresents Announcement posts on the database
    description: >
      This resource represents an annoucement post in the system.
      Each post is identified by a numeric value `id`.
    get:
      tags:
        - Announcement
      description:
        Get annoucement by id.
      responses:
        '200':
          description: View of announcement record
          content:
            application/json:
              schema:
                $ref: ./announcement.json#/definitions/announcementView
        '401':
          description: Non-admin users can only view their own record
        '404':
          description: Not found

    get:
      tags:
        - Announcement
      description:
        Get Announcement by id.
      responses:
        '200':
          description: View of Announcement record
          content:
            application/json:
              schema:
                $ref: ./announcement.json#/definitions/announcementView
        '401':
          description: Non-admin users can only view their own record
        '404':
          description: Not found
    
    post:
      tags:
        - Announcement
      description:
        Create a new Announcement post.
      requestBody:
        description: All new user data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./announcement.json#/definitions/announcement
      responses:
        '201':
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: ./announcement.json#/definitions/announcementUpdated

    put:
      tags:
        - Announcement
      description:
        Update user by id.
      requestBody:
        description: Update a visible property or the password.
        required: true
        content:
          application/json:
            schema:
              $ref: .announcement.json#/definitions/announcementUpdate
      responses:
        '200':
          description: Confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./announcement.json#/definitions/announcementUpdated
        '404':
          description: Not found
...